Common search ast nodes code to a separate file
