+++ arangod/Aql/IResearchViewOptimizerRules.cpp
@@ -30,6 +30,7 @@
 #include "Aql/Expression.h"
 #include "Aql/Function.h"
 #include "Aql/IResearchViewNode.h"
+#include "Aql/LateMaterializedOptimizerRulesCommon.h"
 #include "Aql/Optimizer.h"
 #include "Aql/OptimizerRule.h"
 #include "Aql/Query.h"
@@ -362,132 +363,25 @@ void lateDocumentMaterializationArangoSearchRule(Optimizer* opt,
   }
 }
 
-// ===============================================================================================
 namespace {
-  struct NodeWithAttrs {
-    struct AttributeAndField {
-      std::vector<arangodb::basics::AttributeName> attr;
-      AstNode* astNode;
-      size_t astNodeChildNum;
-      size_t sortFieldNum;
-      std::vector<arangodb::basics::AttributeName> const* sortField;
-    };
-
-    std::vector<AttributeAndField> attrs;
-    CalculationNode* node;
-  };
-
-  bool attributesMatch(IResearchViewSort const& primarySort, NodeWithAttrs& node) {
+  bool attributesMatch(IResearchViewSort const& primarySort, latematerialized::NodeWithAttrs& node) {
     // check all node attributes to be in sort
     for (auto& nodeAttr : node.attrs) {
       size_t sortFieldNum = 0;
       for (auto const& field : primarySort.fields()) {
         if (arangodb::basics::AttributeName::isIdentical(nodeAttr.attr, field, false)) {
-          nodeAttr.sortFieldNum = sortFieldNum;
-          nodeAttr.sortField = &field;
+          nodeAttr.fieldNum = sortFieldNum;
+          nodeAttr.field = &field;
           break;
         }
         ++sortFieldNum;
       }
       // not found
-      if (nodeAttr.sortField == nullptr) {
+      if (nodeAttr.field == nullptr) {
         return false;
       }
     }
     return true;
-  }
-
-  // traverse the AST, using previsitor
-  void traverseReadOnly(AstNode* node, AstNode* parentNode, size_t childNumber,
-                        std::function<bool(AstNode const*, AstNode*, size_t)> const& preVisitor) {
-    if (node == nullptr) {
-      return;
-    }
-
-    if (!preVisitor(node, parentNode, childNumber)) {
-      return;
-    }
-
-    size_t const n = node->numMembers();
-
-    for (size_t i = 0; i < n; ++i) {
-      auto member = node->getMemberUnchecked(i);
-
-      if (member != nullptr) {
-        traverseReadOnly(member, node, i, preVisitor);
-      }
-    }
-  }
-
-  // traversal state
-  struct TraversalState {
-    Variable const* variable;
-    NodeWithAttrs& nodeAttrs;
-    bool optimize;
-    bool wasAccess;
-  };
-
-  // determines attributes referenced in an expression for the specified out variable
-  bool getReferencedAttributes(AstNode* node,
-                               Variable const* variable,
-                               NodeWithAttrs& nodeAttrs) {
-    TraversalState state{variable, nodeAttrs, true, false};
-
-    auto preVisitor = [&state](AstNode const* node,
-        AstNode* parentNode, size_t childNumber) {
-      if (node == nullptr) {
-        return false;
-      }
-
-      switch (node->type) {
-        case NODE_TYPE_ATTRIBUTE_ACCESS:
-          if (!state.wasAccess) {
-            state.nodeAttrs.attrs.emplace_back(
-              NodeWithAttrs::AttributeAndField{std::vector<arangodb::basics::AttributeName>{
-                {std::string(node->getStringValue(), node->getStringLength()), false}}, parentNode, childNumber, 0, nullptr});
-            state.wasAccess = true;
-          } else {
-            state.nodeAttrs.attrs.back().attr.emplace_back(std::string(node->getStringValue(), node->getStringLength()), false);
-          }
-          return true;
-        case NODE_TYPE_REFERENCE: {
-          // reference to a variable
-          auto v = static_cast<Variable const*>(node->getData());
-          if (v == state.variable) {
-            if (!state.wasAccess) {
-              // we haven't seen an attribute access directly before
-              state.optimize = false;
-
-              return false;
-            }
-            std::reverse(state.nodeAttrs.attrs.back().attr.begin(), state.nodeAttrs.attrs.back().attr.end());
-          } else {
-            if (state.wasAccess) {
-              state.nodeAttrs.attrs.pop_back();
-            }
-          }
-          // finish an attribute path
-          state.wasAccess = false;
-          return true;
-        }
-        default:
-          break;
-      }
-
-      if (state.wasAccess) {
-        // not appropriate node type
-        state.wasAccess = false;
-        state.optimize = false;
-
-        return false;
-      }
-
-      return true;
-    };
-
-    traverseReadOnly(node, nullptr, 0, preVisitor);
-
-    return state.optimize;
   }
 }
 
@@ -531,7 +425,7 @@ void lateDocumentMaterializationArangoSearchRule2(Optimizer* opt,
       // without document body usage before that node.
       // this node could be appended with materializer
       bool stopSearch = false;
-      std::vector<NodeWithAttrs> nodesToChange;
+      std::vector<latematerialized::NodeWithAttrs> nodesToChange;
       while (current != loop) {
         switch (current->getType()) {
           case ExecutionNode::SORT:
@@ -542,10 +436,10 @@ void lateDocumentMaterializationArangoSearchRule2(Optimizer* opt,
           case ExecutionNode::CALCULATION: {
             auto calculationNode = ExecutionNode::castTo<CalculationNode*>(current);
             auto astNode = calculationNode->expression()->nodeForModification();
-            NodeWithAttrs node;
+            latematerialized::NodeWithAttrs node;
             node.node = calculationNode;
             // find attributes referenced to view node out variable
-            if (!getReferencedAttributes(astNode, &viewNode->outVariable(), node)) {
+            if (!latematerialized::getReferencedAttributes(astNode, &viewNode->outVariable(), node)) {
               // is not safe for optimization
               stopSearch = true;
             } else if (!node.attrs.empty()) {
@@ -590,7 +484,7 @@ void lateDocumentMaterializationArangoSearchRule2(Optimizer* opt,
         for (auto& node : nodesToChange) {
           std::transform(node.attrs.cbegin(), node.attrs.cend(), std::inserter(uniqueVariables, uniqueVariables.end()),
                          [ast](auto const& attrAndField) {
-                           return std::make_pair(attrAndField.sortField, IResearchViewNode::ViewVariable{attrAndField.sortFieldNum,
+                           return std::make_pair(attrAndField.field, IResearchViewNode::ViewVariable{attrAndField.fieldNum,
                              ast->variables()->createTemporaryVariable()});
                          });
         }
@@ -598,7 +492,7 @@ void lateDocumentMaterializationArangoSearchRule2(Optimizer* opt,
         auto localDocIdTmp = ast->variables()->createTemporaryVariable();
         for (auto& node : nodesToChange) {
           for (auto& attr : node.attrs) {
-            auto it = uniqueVariables.find(attr.sortField);
+            auto it = uniqueVariables.find(attr.field);
             TRI_ASSERT(it != uniqueVariables.cend());
             auto newNode = ast->createNodeReference(it->second.var);
             if (attr.astNode != nullptr) {
